requirements.txt
streamlit
python-dotenv
google-generativeai>=0.5.0
PyPDF2
langchain
langchain-google-genai
langchain-community
chromadb
sentence-transformers
